third_party/WebKit/LayoutTests/external/wpt/preload/single-download-preload-expected.txt

This is a testharness.js-based test.
FAIL Makes sure that preloaded resources are not downloaded again when used assert_equals: resources/CanvasTest.ttf?single-download-preload expected 1 but got 0
Harness: the test ran to completion.
